Automatically load the most recently saved dashboard file
This behavior can be toggled on or off in the application settings
Refactors DashboardTab into its own upper-level class
