+++ api/src/main/java/edu/wpi/first/shuffleboard/api/util/Debouncer.java
@@ -46,4 +46,11 @@ public class Debouncer implements Runnable {
     return debounceDelay;
   }
 
+  /**
+   * Cancels the debouncer. The target will not run unless {@link #run()} is called later.
+   */
+  public void cancel() {
+    future.cancel(true);
+  }
+
 }
