+++ app/src/main/java/edu/wpi/first/shuffleboard/app/Shuffleboard.java
@@ -3,6 +3,7 @@ package edu.wpi.first.shuffleboard.app;
 import edu.wpi.first.shuffleboard.api.sources.recording.Recorder;
 import edu.wpi.first.shuffleboard.api.util.Storage;
 import edu.wpi.first.shuffleboard.app.plugin.PluginLoader;
+import edu.wpi.first.shuffleboard.app.prefs.AppPreferences;
 import edu.wpi.first.shuffleboard.plugin.base.BasePlugin;
 import edu.wpi.first.shuffleboard.plugin.cameraserver.CameraServerPlugin;
 import edu.wpi.first.shuffleboard.plugin.networktables.NetworkTablesPlugin;
@@ -45,8 +46,12 @@ public class Shuffleboard extends Application {
 
   @Override
   public void start(Stage primaryStage) throws IOException {
-    mainPane = FXMLLoader.load(MainWindowController.class.getResource("MainWindow.fxml"));
     onOtherAppStart = () -> Platform.runLater(primaryStage::toFront);
+
+    FXMLLoader loader = new FXMLLoader(MainWindowController.class.getResource("MainWindow.fxml"));
+    mainPane = loader.load();
+    final MainWindowController mainWindowController = loader.getController();
+
     primaryStage.setScene(new Scene(mainPane));
 
     PluginLoader.getDefault().load(new BasePlugin());
@@ -56,6 +61,11 @@ public class Shuffleboard extends Application {
     PluginLoader.getDefault().load(new CameraServerPlugin());
     PluginLoader.getDefault().load(new NetworkTablesPlugin());
     PluginLoader.getDefault().loadAllJarsFromDir(Storage.getPluginPath());
+
+    // Load the most recent save file after loading all plugins
+    if (AppPreferences.getInstance().isAutoLoadLastSaveFile()) {
+      mainWindowController.load(AppPreferences.getInstance().getSaveFile());
+    }
 
     primaryStage.setTitle("Shuffleboard");
     primaryStage.setMinWidth(640);
