+++ app/src/main/java/edu/wpi/first/shuffleboard/app/prefs/AppPreferences.java
@@ -1,16 +1,19 @@
 package edu.wpi.first.shuffleboard.app.prefs;
 
-import com.google.common.annotations.VisibleForTesting;
-import com.google.common.collect.ImmutableList;
-
 import edu.wpi.first.shuffleboard.api.theme.Theme;
 import edu.wpi.first.shuffleboard.api.theme.Themes;
 import edu.wpi.first.shuffleboard.api.util.PreferencesUtils;
 
+import com.google.common.annotations.VisibleForTesting;
+import com.google.common.collect.ImmutableList;
+
+import java.io.File;
 import java.util.prefs.Preferences;
 
+import javafx.beans.property.BooleanProperty;
 import javafx.beans.property.DoubleProperty;
 import javafx.beans.property.Property;
+import javafx.beans.property.SimpleBooleanProperty;
 import javafx.beans.property.SimpleDoubleProperty;
 import javafx.beans.property.SimpleObjectProperty;
 
@@ -22,6 +25,9 @@ public final class AppPreferences {
 
   private final Property<Theme> theme = new SimpleObjectProperty<>(this, "Theme", Themes.INITIAL_THEME);
   private final DoubleProperty defaultTileSize = new SimpleDoubleProperty(this, "defaultTileSize", 128);
+  private final Property<File> saveFile = new SimpleObjectProperty<>(this, "saveFile", null);
+  private final BooleanProperty autoLoadLastSaveFile =
+      new SimpleBooleanProperty(this, "automaticallyLoadLastSaveFile", true);
 
   @VisibleForTesting
   static AppPreferences instance = new AppPreferences();
@@ -33,9 +39,13 @@ public final class AppPreferences {
     Preferences preferences = Preferences.userNodeForPackage(getClass());
     PreferencesUtils.read(theme, preferences, Themes.getDefault()::forName);
     PreferencesUtils.read(defaultTileSize, preferences);
+    PreferencesUtils.read(saveFile, preferences, File::new);
+    PreferencesUtils.read(autoLoadLastSaveFile, preferences);
 
     theme.addListener(__ -> PreferencesUtils.save(theme, preferences, Theme::getName));
     defaultTileSize.addListener(__ -> PreferencesUtils.save(defaultTileSize, preferences));
+    saveFile.addListener(__ -> PreferencesUtils.save(saveFile, preferences, File::getAbsolutePath));
+    autoLoadLastSaveFile.addListener(__ -> PreferencesUtils.save(autoLoadLastSaveFile, preferences));
   }
 
   public static AppPreferences getInstance() {
@@ -48,7 +58,8 @@ public final class AppPreferences {
   public ImmutableList<Property<?>> getProperties() {
     return ImmutableList.of(
         theme,
-        defaultTileSize
+        defaultTileSize,
+        autoLoadLastSaveFile
     );
   }
 
@@ -76,4 +87,27 @@ public final class AppPreferences {
     this.defaultTileSize.set(defaultTileSize);
   }
 
+  public File getSaveFile() {
+    return saveFile.getValue();
+  }
+
+  public Property<File> saveFileProperty() {
+    return saveFile;
+  }
+
+  public void setSaveFile(File saveFile) {
+    this.saveFile.setValue(saveFile);
+  }
+
+  public boolean isAutoLoadLastSaveFile() {
+    return autoLoadLastSaveFile.get();
+  }
+
+  public BooleanProperty autoLoadLastSaveFileProperty() {
+    return autoLoadLastSaveFile;
+  }
+
+  public void setAutoLoadLastSaveFile(boolean autoLoadLastSaveFile) {
+    this.autoLoadLastSaveFile.set(autoLoadLastSaveFile);
+  }
 }
